+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 
 ## Honeybadger 1.9.3 ##
 
+* Filter query strings.
+
+  *Joshua Wood*
+
 * Lock gemspec to faraday ~>0.7 to fix yield discrepancy between 0.7 and
   0.8.